requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-pptx==0.6.23
PyPDF2==3.0.1
python-multipart==0.0.9
pydantic==2.9.0
faiss-cpu==1.8.0
google-generativeai==0.8.0
python-dotenv==1.0.1
langchain-text-splitters==0.3.0
numpy==1.26.4
